Displays original image and image with mask
Shows the image feed from the camera topic (the topic must be changed) and shows a modified image feed that displays only the mask. Must use two arguments (this can be modified as well so arguments are not needed), the first argument is the original image and the second argument is the mask of the original image. This is used to detect the object the mask is placed over.
Ex: python cylinderDetect.py imgName maskImgName